#885 Flattened type hierarchy for CommonEntityDao by removing AbstractEntityDao.
