+++ platform-dao/src/main/java/ua/com/fielden/platform/dao/CommonEntityDao.java
@@ -3,12 +3,15 @@ package ua.com.fielden.platform.dao;
 import static java.lang.String.format;
 import static org.hibernate.LockOptions.UPGRADE;
 import static ua.com.fielden.platform.entity.ActivatableAbstractEntity.ACTIVE;
+import static ua.com.fielden.platform.entity.query.fluent.EntityQueryUtils.cond;
 import static ua.com.fielden.platform.entity.query.fluent.EntityQueryUtils.fetchAggregates;
 import static ua.com.fielden.platform.entity.query.fluent.EntityQueryUtils.from;
+import static ua.com.fielden.platform.entity.query.fluent.EntityQueryUtils.orderBy;
 import static ua.com.fielden.platform.entity.query.fluent.EntityQueryUtils.select;
 import static ua.com.fielden.platform.types.tuples.T2.t2;
 
 import java.io.IOException;
+import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -17,6 +20,7 @@ import java.util.Collections;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -41,13 +45,17 @@ import ua.com.fielden.platform.entity.AbstractEntity;
 import ua.com.fielden.platform.entity.AbstractPersistentEntity;
 import ua.com.fielden.platform.entity.AbstractUnionEntity;
 import ua.com.fielden.platform.entity.ActivatableAbstractEntity;
+import ua.com.fielden.platform.entity.DynamicEntityKey;
 import ua.com.fielden.platform.entity.IContinuationData;
 import ua.com.fielden.platform.entity.annotation.CompanionObject;
 import ua.com.fielden.platform.entity.annotation.DeactivatableDependencies;
+import ua.com.fielden.platform.entity.annotation.EntityType;
 import ua.com.fielden.platform.entity.annotation.MapTo;
 import ua.com.fielden.platform.entity.annotation.Required;
+import ua.com.fielden.platform.entity.factory.EntityFactory;
 import ua.com.fielden.platform.entity.factory.ICompanionObjectFinder;
 import ua.com.fielden.platform.entity.fetch.FetchModelReconstructor;
+import ua.com.fielden.platform.entity.fetch.IFetchProvider;
 import ua.com.fielden.platform.entity.meta.MetaProperty;
 import ua.com.fielden.platform.entity.meta.PropertyDescriptor;
 import ua.com.fielden.platform.entity.query.EntityAggregates;
@@ -55,9 +63,13 @@ import ua.com.fielden.platform.entity.query.EntityFetcher;
 import ua.com.fielden.platform.entity.query.IFilter;
 import ua.com.fielden.platform.entity.query.IdOnlyProxiedEntityTypeCache;
 import ua.com.fielden.platform.entity.query.QueryExecutionContext;
+import ua.com.fielden.platform.entity.query.fluent.EntityQueryProgressiveInterfaces.ICompoundCondition0;
+import ua.com.fielden.platform.entity.query.fluent.EntityQueryProgressiveInterfaces.IWhere0;
 import ua.com.fielden.platform.entity.query.fluent.fetch;
 import ua.com.fielden.platform.entity.query.model.AggregatedResultQueryModel;
+import ua.com.fielden.platform.entity.query.model.ConditionModel;
 import ua.com.fielden.platform.entity.query.model.EntityResultQueryModel;
+import ua.com.fielden.platform.entity.query.model.OrderingModel;
 import ua.com.fielden.platform.entity.query.model.QueryModel;
 import ua.com.fielden.platform.error.Result;
 import ua.com.fielden.platform.file_reports.WorkbookExporter;
@@ -89,7 +101,7 @@ import ua.com.fielden.platform.utils.Validators;
  * @param <K>
  *            -- entitie's key type
  */
-public abstract class CommonEntityDao<T extends AbstractEntity<?>> extends AbstractEntityDao<T> implements ISessionEnabled {
+public abstract class CommonEntityDao<T extends AbstractEntity<?>> implements IEntityDao<T>, ISessionEnabled {
 
     private static final String ERR_MSG_NO_QUERY_PROVIDED = "There was no query provided to retrieve the data.";
     
@@ -125,6 +137,36 @@ public abstract class CommonEntityDao<T extends AbstractEntity<?>> extends Abstr
     private final boolean hasSaveOverridden;
 
     private boolean instrumented = true;
+
+    private final Class<? extends Comparable<?>> keyType;
+    private final Class<T> entityType;
+    private IFetchProvider<T> fetchProvider;
+    
+    @Inject
+    private EntityFactory entityFactory;
+
+    /**
+     * The default constructor, which looks for annotation {@link EntityType} to identify the entity type automatically.
+     * An exception is thrown if the annotation is missing. 
+     *
+     * @param entityType
+     */
+    protected CommonEntityDao(final IFilter filter) {
+        final EntityType annotation = AnnotationReflector.getAnnotation(getClass(), EntityType.class);
+        if (annotation == null) {
+            throw new EntityCompanionException(format("Companion object [%s] is missing @EntityType annotation.", getClass().getName()));
+        }
+        this.entityType = (Class<T>) annotation.value();
+        this.keyType = AnnotationReflector.getKeyType(entityType);
+        
+        this.filter = filter;
+        this.hasSaveOverridden = isSaveOverridden();
+        this.deleteOps = new CommonEntityCompanionDeleteOperations<>(this);
+    }
+
+    protected boolean getFilterable() {
+        return false;
+    }
     
     @Override
     public boolean instrumented() {
@@ -141,17 +183,6 @@ public abstract class CommonEntityDao<T extends AbstractEntity<?>> extends Abstr
         final CommonEntityDao<T> co = (CommonEntityDao<T>) injector.getInstance(coType);
         co.instrumented = false;
         return (E) co;
-    }
-    
-    /**
-     * A principle constructor.
-     *
-     * @param entityType
-     */
-    protected CommonEntityDao(final IFilter filter) {
-        this.filter = filter;
-        this.hasSaveOverridden = isSaveOverridden();
-        this.deleteOps = new CommonEntityCompanionDeleteOperations<T>(this);
     }
     
     private boolean isSaveOverridden() {
@@ -194,7 +225,7 @@ public abstract class CommonEntityDao<T extends AbstractEntity<?>> extends Abstr
         try {
             sess.cancelQuery();
         } catch (final Exception ex) {
-            ex.printStackTrace();
+            logger.warn(ex);
         }
         return true;
     }
@@ -214,22 +245,34 @@ public abstract class CommonEntityDao<T extends AbstractEntity<?>> extends Abstr
     }
 
     @Override
+    public T findById(final Long id) {
+        return findById(id, null);
+    }
+
+    @Override
     @SessionRequired
     public T findById(final Long id, final fetch<T> fetchModel) {
-        return super.findById(id, fetchModel);
+        return fetchOneEntityInstance(id, fetchModel);
     }
 
     @Override
     @SessionRequired
     public T findByKeyAndFetch(final fetch<T> fetchModel, final Object... keyValues) {
-        return super.findByKeyAndFetch(fetchModel, keyValues);
+        try {
+            return getEntity(instrumented() ? from((createQueryByKey(keyValues))).with(fetchModel).model() : from((createQueryByKey(keyValues))).with(fetchModel).lightweight().model());
+        } catch (final EntityCompanionException e) {
+            throw e;
+        } catch (final Exception e) {
+            throw new IllegalStateException(e);
+        }
     }
 
     @Override
     @SessionRequired
     public T findByKey(final Object... keyValues) {
-        return super.findByKey(keyValues);
+        return findByKeyAndFetch(null, keyValues);
     }
+
 
     /**
      * {@inheritDoc} 
@@ -1366,6 +1409,195 @@ public abstract class CommonEntityDao<T extends AbstractEntity<?>> extends Abstr
     @SessionRequired
     protected <E extends AbstractEntity<?>> int defaultBatchDeleteByPropertyValues(final String propName, final List<E> propEntities) {
         return deleteOps.defaultBatchDeleteByPropertyValues(propName, propEntities);
+    }
+
+    protected QueryExecutionModel<T, EntityResultQueryModel<T>> produceDefaultQueryExecutionModel(final Class<T> entityType) {
+        final EntityResultQueryModel<T> query = select(entityType).model();
+        query.setFilterable(getFilterable());
+        final OrderingModel orderBy = orderBy().prop(AbstractEntity.ID).asc().model();
+        return instrumented() ? from(query).with(orderBy).model() : from(query).with(orderBy).lightweight().model();
+    }
+
+    protected QueryExecutionModel<T, EntityResultQueryModel<T>> getDefaultQueryExecutionModel() {
+        return produceDefaultQueryExecutionModel(entityType);
+    }
+
+    @Override
+    public Class<T> getEntityType() {
+        return entityType;
+    }
+
+    @Override
+    public Class<? extends Comparable<?>> getKeyType() {
+        return keyType;
+    }
+
+    private T fetchOneEntityInstance(final Long id, final fetch<T> fetchModel) {
+        try {
+            final EntityResultQueryModel<T> query = select(getEntityType()).where().prop(AbstractEntity.ID).eq().val(id).model();
+            query.setFilterable(getFilterable());
+            return getEntity(instrumented() ? from(query).with(fetchModel).model(): from(query).with(fetchModel).lightweight().model());
+        } catch (final Exception e) {
+            throw new IllegalStateException(e);
+        }
+    }
+
+    /**
+     * Method checks whether the key of the entity type associated with this DAO if composite or not.
+     *
+     * If composite then <code>WHERE</code> statement is build using composite key members and the passed values. The number of values should match the number of composite key
+     * members.
+     *
+     * Otherwise, <code>WHERE</code> statement is build using only property <code>key</code>.
+     *
+     * The created query expects a unique result, and throws a runtime exception if this is not the case.
+     *
+     * TODO Need to consider the case of polymorphic associations such as Rotable, which can be both Bogie and/or Wheelset.
+     */
+    @Override
+    public boolean entityWithKeyExists(final Object... keyValues) {
+        final T entity = findByKeyAndFetch(null, keyValues);
+        return entity != null;
+    }
+
+    @Override
+    public T findByEntityAndFetch(final fetch<T> fetchModel, final T entity) {
+        if (entity.getId() != null) {
+            return findById(entity.getId(), fetchModel);
+        } else {
+            return findByKeyAndFetch(fetchModel, entity.getKey());
+        }
+    }
+
+    /**
+     * Convenient method for composing a query to select an entity by key value.
+     *
+     * @param keyValues
+     * @return
+     */
+    protected EntityResultQueryModel<T> createQueryByKey(final Object... keyValues) {
+        if (keyValues == null || keyValues.length == 0) {
+            throw new IllegalArgumentException("No key values provided.");
+        }
+        final EntityResultQueryModel<T> query = attachKeyConditions(select(getEntityType()).where(), keyValues).model();
+        query.setFilterable(getFilterable());
+        return query;
+    }
+    
+    /**
+     * Creates a query for entities by their keys (simple or composite). If <code>entitiesWithKeys</code> are empty -- returns empty optional. 
+     * 
+     * @param entityType -- the entity type
+     * @param entitiesWithKeys -- the entities with <b>all</b> key values correctly fetched / assigned
+     * @return
+     */
+    public Optional<EntityResultQueryModel<T>> createQueryByKeyFor(final Collection<T> entitiesWithKeys) {
+        IWhere0<T> partQ = select(getEntityType()).where();
+        final List<Field> keyMembers = Finder.getKeyMembers(getEntityType());
+        
+        for (final Iterator<T> iter = entitiesWithKeys.iterator(); iter.hasNext();) {
+            final T entityWithKey = iter.next();
+            final ICompoundCondition0<T> or = attachKeyConditions(partQ, keyMembers, keyMembers.stream().map(keyMember -> entityWithKey.get(keyMember.getName())).toArray());
+            if (iter.hasNext()) {
+                partQ = or.or();
+            } else {
+                return Optional.of(or.model());
+            }
+        }
+
+        return Optional.empty();
+    }
+    
+    /**
+     * Attaches key member conditions to the partially constructed query <code>entryPoint</code> based on its values. 
+     * 
+     * @param entryPoint
+     * @param keyValues
+     * @return
+     */
+    protected ICompoundCondition0<T> attachKeyConditions(final IWhere0<T> entryPoint, final Object... keyValues) {
+        return attachKeyConditions(entryPoint, Finder.getKeyMembers(getEntityType()), keyValues);
+    }
+
+    /**
+     * Attaches key member conditions to the partially constructed query <code>entryPoint</code> based on its values. 
+     * 
+     * @param entryPoint
+     * @param keyMembers
+     * @param keyValues
+     * @return
+     */
+    protected ICompoundCondition0<T> attachKeyConditions(final IWhere0<T> entryPoint, final List<Field> keyMembers, final Object... keyValues) {
+        if (getKeyType() == DynamicEntityKey.class) {
+            // let's be smart about the key values and support the case where an instance of DynamicEntityKey is passed.
+            final Object[] realKeyValues;
+            if (keyValues.length == 1 && keyValues[0] instanceof DynamicEntityKey) {
+                realKeyValues = ((DynamicEntityKey) keyValues[0]).getKeyValues(); 
+            } else {
+                realKeyValues = keyValues;
+            }
+
+            if (keyMembers.size() != realKeyValues.length) {
+                throw new EntityCompanionException(format("The number of provided values (%s) does not match the number of properties in the entity composite key (%s).", realKeyValues.length, keyMembers.size()));
+            }
+
+            ICompoundCondition0<T> cc = entryPoint.condition(buildConditionForKeyMember(keyMembers.get(0).getName(), keyMembers.get(0).getType(), realKeyValues[0]));
+
+            for (int index = 1; index < keyMembers.size(); index++) {
+                cc = cc.and().condition(buildConditionForKeyMember(keyMembers.get(index).getName(), keyMembers.get(index).getType(), realKeyValues[index]));
+            }
+            return cc;
+        } else if (keyValues.length != 1) {
+            throw new EntityCompanionException(format("Only one key value is expected instead of %s when looking for an entity by a non-composite key.", keyValues.length));
+        } else {
+            return entryPoint.condition(buildConditionForKeyMember(AbstractEntity.KEY, getKeyType(), keyValues[0]));
+        }
+    }
+
+    private ConditionModel buildConditionForKeyMember(final String propName, final Class<?> propType, final Object propValue) {
+        if (propValue == null) {
+            return cond().prop(propName).isNull().model();
+        } else if (String.class.equals(propType)) {
+            return cond().lowerCase().prop(propName).eq().lowerCase().val(propValue).model();
+        } else if (Class.class.equals(propType)) {
+            return cond().prop(propName).eq().val(((Class<?>) propValue).getName()).model();
+        } else {
+            return cond().prop(propName).eq().val(propValue).model();
+        }
+    }
+
+    public final IFetchProvider<T> getFetchProvider() {
+        if (fetchProvider == null) {
+            fetchProvider = createFetchProvider();
+        }
+        return fetchProvider;
+    }
+
+    /**
+     * Creates fetch provider for this entity companion.
+     * <p>
+     * Should be overridden to provide custom fetch provider.
+     *
+     * @return
+     */
+    protected IFetchProvider<T> createFetchProvider() {
+        // provides a very minimalistic version of fetch provider by default (only id and version are included)
+        return EntityUtils.fetch(getEntityType());
+    }
+    
+    protected EntityFactory getEntityFactory() {
+        return entityFactory;
+    }
+ 
+    /**
+     * Instantiates an instrumented new entity of the type for which this object is a companion.
+     * The default entity constructor, which should be protected, is used for instantiation.
+     *
+     * @return
+     */
+    @Override
+    public T new_() {
+        return entityFactory.newEntity(getEntityType());
     }
 
 }
